Two-part tariff review screens - Adding purpose to element
https://eaflood.atlassian.net/browse/WATER-4513

During testing for the two-part tariff review screens by the billing and data team, it was noted that adding the purpose to the end of the charge elements description, along with abstraction period dates, would simplify the review process. This PR aims to include the charge elements purpose text on the licence review page.