Add ability to search by Licence Number in 2 Part Tariff Review https://eaflood.atlassian.net/browse/WATER-4515
The pagination can make it difficult to find a specific licence in the review so the business has requested the ability to search by Licence Number.

Currently the user is only able to search by Licence Holder Name. This change is going to allow a search on either Licence Holder Name or Licence Number from the same text field (similar to the main search bar).