Remove multiple issues showing as review status on two-part tariff pages
https://eaflood.atlassian.net/browse/WATER-4517

During testing for the two-part tariff review screens by the billing and data team, it was noted that if a licence had multiple issues (even if the issues where all of a ready status) that the licence would be marked as a review status. The point of the 'statuses' was to make issues visible to users and direct them if there was action needed. If a licence has two issues and these are both of a 'ready' status then it doesn't need any action from the team. This PR is to change the logic to remove multiple issues being flagged as a review status.