Remove multiple issues showing as review status on two-part tariff pages (#1097)
https://eaflood.atlassian.net/browse/WATER-4517

During testing for the two-part tariff review screens by the billing and data team, it was noted that if a licence had multiple issues (even if they were all of a 'ready' status) the licence would be marked as a 'review' status. The point of the statuses was to make issues visible to users and direct them if action was needed. If a licence has two issues, both of a 'ready' status, then it doesn't have any action for the team to complete. This PR changes the logic to remove multiple issues flagging as a review status.