Fix pagination text in 2PT review screen https://eaflood.atlassian.net/browse/WATER-4515
During testing it has been found that in the two-part tariff review screen. If a filter is applied to the results resulting in no records being returned, the title is displayed as "Review licences (page 1 of 0)". The pagination is incorrectly being displayed in the title. The pagination should only be displayed if more than 1 page of data exists.

This issue is caused by the logic used to determine if the pagination should be displayed. Currently it is not displayed if the number of pages equals 1. This needs to be updated so that it is not displayed if the number of pages is less than 2.

This PR will fix that issue.