Link licences to licence agreements (#1098)
https://eaflood.atlassian.net/browse/WATER-4512

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to generate return requirements from the licence's abstraction data. On the `/setup` page, they can select this option and the service is expected to generate return requirements using the source abstraction data against the licence. The other options are to copy existing return requirements or set them up manually.

This change is in support of that. We need to get the licence's agreements to inform how we generate the return requirements. But we've spotted that though we have both `LicenceModel` and `LicenceAgreementModel` no [Objection.js relationship](https://vincit.github.io/objection.js/guide/relations.html#relations) has been added to link them.