Fix licence 2 licence gauging station relationship
https://eaflood.atlassian.net/browse/WATER-4331

The `LicenceGaugingStationModel` was in [Display monitoring stations in view licence page summary tab](https://github.com/DEFRA/water-abstraction-system/pull/773) as part of our work to replace the legacy view licence page.

`water.licence_gauging_station` is simply the many-to-many table that links licences to gauging stations. When we get those we'll typically set the relationship up, for example, that `licences` links to `gauging_stations` _through_ `licence_gauging_stations`.

That was what we did in this case. Only, we named the relationship `licenceGaugingStations`. This is incorrect because if we did do an [Objection.js withGraphFetched()](https://vincit.github.io/objection.js/api/query-builder/eager-methods.html#withgraphfetched) we'd get `GaugingStationModel`s back, _not_ `LicenceGaugingStationModel`s.

Also, we never added any tests for the relationship!

This change fixes the name of the relationship and adds the missing tests.