Update two-part tariff review pages to add Cypress data test attributes (#1105)
https://eaflood.atlassian.net/browse/WATER-4450

We've finished building all the pages for the two-part tariff review and are now working on the acceptance tests. While writing these tests, we found a few places where the data test attribute on the view needs to be updated to make it clearer and easier to find in the acceptance tests file.