Fix view licence summary presenter tests
We needed to update this because of the other changes we've made. But when we opened up the tests we found they didn't conform to our normal pattern of presenter tests.

So, we've ensured we have one test that asserts the full response. Then we've replaced any tests that were based on properties going in to be focused on the properties coming out.