Add requirements for returns view
Ticket - https://eaflood.atlassian.net/browse/WATER-4422

Adds a new page to allow users to view the requirements for returns.

A user can click a view link on the licence setup tab under the requirements for returns.

This page is similar to previous work done on the requirements for returns check page.