Fix licence 2 licence gauging station relationship (#1102)
https://eaflood.atlassian.net/browse/WATER-4331

** Where we started!

The `LicenceGaugingStationModel` was added in [Display monitoring stations in view licence page summary tab](https://github.com/DEFRA/water-abstraction-system/pull/773) as part of our work to replace the legacy view licence page.

`water.licence_gauging_stations` links licences to gauging stations. When we get these, we'll typically setup the relationship, for example, as `licences` links to `gauging_stations` _through_ `licence_gauging_stations`.

That was what we did in this case. Only we named the relationship `licenceGaugingStations`. This is incorrect because if we did do an [Objection.js withGraphFetched()](https://vincit.github.io/objection.js/api/query-builder/eager-methods.html#withgraphfetched) we'd get `GaugingStationModel`s back, _not_ `LicenceGaugingStationModel`s.

But in this case, `water.licence_gauging_stations` also contains details about the abstraction periods that apply and the certain metrics (though, to be honest, we don't know why we record them!) We'll want that information as well at some point, so a _through_ relationship is not really appropriate.

Also, we never added any tests for the relationship!

** Where we finished!

Digging into this relationship and correcting it then meant needing to make changes to `FetchLicenceSummaryService`. But that brought to light that the fetch query was doing things in a way that could be simplified with knock-on effects for the presenter.

Then in the presenter tests, we saw that there wasn't a test that covered a full response, and we'd split them based on the properties passed in and not what we get out.

We also found that a number of these models didn't have unit tests themselves. Finally, the migration scripts didn't match up to what is actually in the real DB.

SO, this became a major tidy up of fetching and presenting a licence for the view summary. There is still more we think can be done. But we've drawn a line at this point 😁.