Fix model migrations since last release
> Warning! This will break local environments unless you manually remove existing tables and views in the `public` schema.

We've been doing work for both return requirements and two-part tariff that has resulted in a number new migrations being added and existing ones being edited.

However, because these 2 features are large, complex and constantly in flux we've often needed to make changes. Or, we simply got it wrong and the original migrations was incorrect.

We've just spotted that a recent attempt to fix a migration actually ended up reverting an earlier fix! We've always had a policy that migrations are not 'fixed' until they are shipped to production. So, we're taking a pause and tidying up the currently 'unshipped' migrations.

This means for example, where we have created a migration, then altered it a number of times that those alterations will be added to the original `create` then deleted.

We'll also take the opportunity to get the files names back on track with what they should be.