Add Return Requirement helpers to data loader
https://eaflood.atlassian.net/browse/WATER-4283

When we added the return requirement models as part of our work to switch managing them in NALD to WRLS, we forgot to add them to the data load service, which we use for acceptance testing. Makes it hard to write acceptance tests for return requirements! 🤦

This change corrects that omission.