Fix missing view licence returns status tags
https://eaflood.atlassian.net/browse/WATER-4444

We are replacing the legacy view licence page with a new one. That means we needed to recreate the returns tab, though with some improvements to the display.

However, further testing has shown we've missed some of the possible status, specifically `void` and `received`. Plus we were automatically converting `due` to `overdue` when this should only be if the due date has been passed.

This change fixes the logic for the return statuses. Whilst we are at it we add a new macro for handling display the tag to keep things consistent with how we show tags in other areas of the app.