Use abstraction data to create return requirements (#1107)
https://eaflood.atlassian.net/browse/WATER-4442

> Part of the work to replace NALD for handling return requirements

We want to be able to offer users the option to generate return requirements from the licence's abstraction data. On the `/setup` page, they can select this option and the service is expected to generate return requirements using the source abstraction data against the licence. The other options are to copy existing return requirements or set them up manually.

This change is about implementing the logic behind the scenes when 'Start by using abstraction data' is selected. We fetch the licence's abstraction data, generate return requirements from it, and save them to the session before redirecting the user to the `/check` page.