Update validation msg for ret. req. copy existing
https://eaflood.atlassian.net/browse/WATER-4283

> Part of the work to replace NALD for handling return requirements

In [Complete Select existing return requirement page](https://github.com/DEFRA/water-abstraction-system/pull/895) we added the controls and logic to enable a user to copy the return requirements from an existing return version into the session.

The H1 title we used though was though to be wrong. So, we [Correct copy existing return req. page title](https://github.com/DEFRA/water-abstraction-system/pull/1092). When we did that, we also updated the validation message as all our current messages just replay the H1 title.

Again however, this was thought to be incorrect. Our UAT team have decided the message we should show when an option is not selected is **Select a previous requirements for returns**.

This is that change.