Temporary session storage cleanup job (#1100)
https://eaflood.atlassian.net/browse/WATER-4281

As part of the work on returns, we have had to build in support for temporary session storage into the `water-abstraction-system`. This is what allows us to track data across pages as part of a journey until it’s ready to be properly persisted.

When a user gets to the end of their journey they can explicitly complete or cancel it. Either way, we’ll delete our temporary session record when they do this. But if they abandon the journey without doing this we’ll be left with the temporary record. Over time this would build up and affect the performance of any pages that use this temporary storage.

This PR will create a job that can be triggered through an API call that deletes any record in sessions where the created on date is more than 24 hours ago.