Add load helpers for copy existing acceptance test
https://eaflood.atlassian.net/browse/WATER-4488

When writing acceptance tests for the copy from existing requirement journey, load helper entities are requested by the acceptance
tests to populate the DB. This creates associated records in the DB needed for the copy from existing acceptance tests.

This PR focuses on adding the missing load helpers needed for the return requirements journey in order to create the copy from
existing requirements acceptance test.