Spike for two-part tariff annual billing
https://eaflood.atlassian.net/browse/WATER-4383

We've built the first part of the two-part tariff annual billing process; matching and allocating. Once users have reviewed the results and made any changes we need to generate the actual bills.

We'e previously done that with SROC annual and supplementary and we know we can make use of a lot of it. But two-part tariff is different in that we already know which licences to include and key details need to come from the review results not the charging information.

So, this is a spike; hack together a 'working' billing engine that generates the bills using the review data.

We never intend to merge this change. But we'll use it to direct and inform the changes we do need to make.