Update validation msg for ret. req. copy existing
https://eaflood.atlassian.net/browse/WATER-4283

In [Update validation msg for ret. req. copy existing](https://github.com/DEFRA/water-abstraction-system/pull/1112) we updated the validation message if a user doesn't select an option in the copy existing return requirements page to "Select a previous requirements for returns".

Our design team however were originally going to use "Select a return version". Having seen the updated text it in the flesh, they've decided to go with their original plan.

This change sets the validation message to be "Select a return version".