Add created_by to return versions and link models (#1126)
https://eaflood.atlassian.net/browse/WATER-4522

> Part of the work to replace NALD for handling return requirements

When we take control of the setup of return requirements, we'll be persisting them in the same tables that they are currently imported into.

One omission is that there is no means to identify who created the return version. Admittedly, there is nothing in NALD to indicate this, so there is nothing to import. But we intend to capture this information when we start persisting the records.

We've [added the field to the table](https://github.com/DEFRA/water-abstraction-service/pull/2567). This change makes the field available in the model, and links the return version and user models together using it.