Add feature flag to enable the system licence view
There are currently links directing the pages to the legacy licence view pages, built previously in the water-abstraction-ui repo. This service is slowly being replaced.

We have rebuilt these pages (with some tweaks) in this repo with the intention of using the water-abstraction-system to view and manage licences.

This change will allow us to navigate to the water-abstraction-system licence view pages instead of the 'legacy' licence view pages in a controlled manner.

This change will be mirrored in water-abstraction-ui.

https://eaflood.atlassian.net/browse/WATER-4466