Add new two-part tariff generate bill run endpoint (#1123)
https://eaflood.atlassian.net/browse/WATER-4196

> Part of the work for two-part tariff annual billing

We're ready to generate a bill run from our two-part tariff review data and with [Add Continue bill run btn to 2PT review screen](https://github.com/DEFRA/water-abstraction-system/pull/1122) we have the means to trigger it.

Next, we need the endpoint that the button will hit. This change adds the endpoint (route and controller) plus a 'shell' `GenerateBillRunService` we can start to develop in subsequent changes.