Add sort by to data fetched for two-part tariff review
During the writing of the acceptance tests for the two-part tariff review pages, we realised there was no ordering on the data being pulled out and displayed. The review licence page fetched the charge versions, charge references, charge elements and returns for a two-part tariff licence. Some licences can have multiple or all of these. With no ordering happening on the review pages, this makes the acceptance tests difficult to right as they pass for one persons local environment but fails on another persons because they are being pulled out in different orders. This PR is to add some ordering on the pages.