Add orderBy to data fetched for two-part tariff review  (#1131)
https://eaflood.atlassian.net/browse/WATER-4450

During the writing of the acceptance tests for the two-part tariff review pages, we realised there was no ordering on the data being pulled out and displayed. The review licence page fetched the charge versions, charge references, charge elements and returns for a two-part tariff licence. Some licences can have multiple or all of these. With no order happening on the review pages, this makes the acceptance tests difficult to write as they pass for one person's local environment but fail on another because they are being pulled out in different orders. This PR is to add some ordering on the pages.

The charge versions are being ordered by their start dates.
The charge references are ordered by the highest subsistence charge
The charge elements are being ordered by the highest authorised volume
The returns are being ordered by their start dates.