Feature-requirements-for-returns-view-page (#1118)
* Add requirements for returns view page

https://eaflood.atlassian.net/browse/WATER-4422

The requirements for return journey has been built and the check page has been finished. This showed a page with requirements for returns allowing users to make changes to requirements for returns.

This piece of work will replicate the requirements for returns check page but will not allow users to make changes (the card and information is identical, apart from the title and a return reference).

There will be a link "View" on the licence set up page that links to this page. That is added as part of this change. Work for this was done here - https://github.com/DEFRA/water-abstraction-system/pull/1054