Fix ordering of points and purposes in ret. req.
https://eaflood.atlassian.net/browse/WATER-4478

When running the acceptance tests we've created for the returns requirement setup journeys we've encountered random failures. Typically it will pass on one machine but then fail on another.

We've tracked the root cause down to the fact we have not imposed an order in how we list the purposes and points during the journey. This means the test is always selecting option 1, for example, and then asserting it is 'Foo'. But depending on the machine the test is running on option 1 might have displayed as 'Bar'. This means the assertion fails.

This change ensures the points and purposes in the journey are listed in a consistent controlled manor. This will fix the tests and make things better for all!