Fix default licence view createdBy
https://eaflood.atlassian.net/jira/software/c/projects/WATER/boards/96?selectedIssue=WATER-4422

When a return version does not have a user linked to the record the default string should be "Migrated from NALD".

The assumption is a user can not create a return unless they are logged in. When the return is saved the user will be captured and stored in the createBty column. Therefore, if the created by field is null then it is safe to assume the data has been migrated from NALD.