Add company address model and relationships
https://eaflood.atlassian.net/browse/WATER-3981

We recently [added a new acceptance test data loader](https://eaflood.atlassian.net/browse/WATER-3981) to replace the slow and flaky legacy acceptance test data loader in [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service).

However a couple of our acceptance tests fail when using our new test data. After investigation it turns out that the legacy charge information setup journey relies on their being a `crm_v2.company_addresses` record linking a company to its addresses.

It can find the companies linked to our licence. But once you select the company to use it then goes looking for the possible addresses. With the old test data you get to pick from 2 because there are `crm_v2.company_addresses`. With our new test data you are taken to the enter a postcode screen as a fallback to the legacy code not finding any addresses.

At the moment this is the only reason we've come across that depends on this table. But we need it to make this journey work.

SO, this change adds the model and links it to `CompanyModel`. We then ensure it is cleaned down in the tear-down and available in our data loader.