Fix licence set up links
During a refactor the links for the page where wrapped up into a single object.

The previous links where missed. This change adds the link to the href block of the gov uk buttons.

The links are used in the if condition and work / where done as part of the original refactor.