Update two-part tariff review pages to add Cypress data test attributes (#1139)
https://eaflood.atlassian.net/browse/WATER-4450

During the work to change the licence view page from the legacy code to the new code, there was a change to any links that directed a user to view a licence. Since the licence review page for our two-part tariff work has 3 links to view the licence, this page was altered in those changes. When those changes were made the data-test attributes for the 3 links were removed and subsequently the acceptance tests began failing.
This PR is to add those back in.