Fix company contact model relationships
https://eaflood.atlassian.net/browse/WATER-4434

We added `CompanyContactModel` as part of our work to replace the legacy view licence page with a new one. It is needed to support the 'Contacts' tab.

At the time we added the model we failed to realise that the relationships are set as `HasManyRelation` when they should be `BelongsToOne`. We also didn't spot that `role_id` in the legacy table points to `crm_v2.roles` which we have named `LicenceRoles` in our views to different the various `roles` tables across the legacy schemas.

So, (sorry folks!) we need to go back and correct a migration (we can do this because it has not yet been pushed to production).

Finally, we complete _all_ the relationships for the model so we have a complete picture.