Enhance the padding-line lint rules
https://github.com/DEFRA/water-abstraction-team/issues/115

We enhanced our linting rules to enforce [spaces around blocks and expressions](https://github.com/DEFRA/water-abstraction-system/pull/1143). This is a great step to helping folks 'fall into the pit of success' and avoid reviewers having to nit-pick PRs.

But having enabled it we've spotted it is highlighting some things that are perfectly acceptable. We also think there are a few other options we could add.

So, this change is about building on the [padding-line-between-statements](https://eslint.style/rules/js/padding-line-between-statements) we've already added.