Enhance the padding-line lint rules (#1146)
https://github.com/DEFRA/water-abstraction-team/issues/115

We enhanced our linting rules to enforce [spaces around blocks and expressions](https://github.com/DEFRA/water-abstraction-system/pull/1143). This is a great step to helping folks 'fall into the pit of success' and avoid reviewers having to nitpick PRs.

But having enabled it, we've spotted that it is highlighting some things that are perfectly acceptable.

We also think there are a few other options we could add.

So, this change is about building on the [padding-line-between-statements](https://eslint.style/rules/js/padding-line-between-statements) we've already added.

**Notes**

- Remove the config that is causing the violation
- Ensure blanks after 'use strict' declaration
- Ensure blanks after variable declarations

That last change ensures we put a blank line between variable declarations and the logic.

Running the linter across all files highlighted a number of violations. But the first 3 we fixed in this change give a good example of what to expect from this rule.

- `app/presenters/base.presenter.js` - simple example of separating variables from logic to help readability
- `app/presenters/bill-runs/two-part-tariff/review-bill-run.presenter.js` - demonstrating that comments are handled!
- `app/presenters/licences/set-up.presenter.js` - this is an example of compromise. It could be rightly argued it is better to have the var declaration next to the logic. But for the sake of consistency and having a tool to automate this stuff, we'll compromise and make the change