Fetch conditions from licence summary service
https://eaflood.atlassian.net/browse/WATER-4322

> Part of the work to replace the legacy view licence page

The `ViewLicenceSummaryService` is calling `FetchLicenceAbstractionConditionsService` and `FetchLicenceSummaryService` and then passing both results to the presenter.

Having checked, it is the only service using `FetchLicenceAbstractionConditionsService` and we don't really need the data preparation it is doing.

If the relationship was added between `LicenceVersionPurposeModel` and `LicenceVersionPurposeConditionModel` we could get everything we need from the `FetchLicenceSummaryService`!

So, this change makes that so.