Enhance view licence page contacts tab
https://eaflood.atlassian.net/browse/WATER-4433
https://eaflood.atlassian.net/browse/WATER-4434

> Part of the work to replace the legacy view licence page

Having now seen the updated view licence page's "Contacts" tab we think we can improve the design a little. The link `Go to customer contacts` should be displayed in the **Customer contacts** section rather than at the top. And we should display both sections all the time, though with the default _No [type] contacts found._ when applicable.

This change updates the page with these changes.