Fix view licence abs amounts issue plus refactor (#1132)
https://eaflood.atlassian.net/browse/WATER-4322

> Part of our work to replace the legacy view licence page

We found our new licence was crashing when we tried to view a specific licence. When we dug into it, the problem was that even though it had a current version in the `PermitLicenceModel`s `licenceDataValue` field, and that version had purposes, those purposes didn't have abstraction amounts.

It's not common for licences to be in this state, but there is not much we can do as the data comes from NALD. It would have to be corrected there.

So, we need to update the `ViewLicenceSummaryPresenter` code to handle this. But it has been some time since we built the summary tab so we have more experience with the licence abstraction data. For example, we now know these values can be taken from somewhere other than the JSON blob the previous team was using.

This means, along with fixing the issue, we intend to do a little 'housekeeping' along the way to try and simplify how things are working, and to use normalised data wherever we can.