Refactor view to use govuk table component
The key change is we can make use of the GOV.UK table component rather than hand-cranking our own. This should mean if changes are made to the component we'll automatically see them.

Along with the refactoring we move the customer contacts link to be part of the customer contacts caption and we move to a design that always displays both sections, even when there are no licence or customer contacts.