Fix duplicate monitoring stations in view licence
https://eaflood.atlassian.net/browse/WATER-4331

> Part of the work to replace the legacy view licence page

After attempting to [simplify the logic that builds the view licence page's summary tab](https://github.com/DEFRA/water-abstraction-system/pull/1132) we've spotted that we may have introduced a bug.

We have found a licence that has been linked to the same gauging station multiple times. This results in seeing the same station listed multiple times in the summary.

Like the other sections, where something is duplicated we should see only one entry. This change fixes the logic and ensures the list is distinct.