Display msg for multiple abs amounts in view
https://eaflood.atlassian.net/browse/WATER-4333

> Part of the work to replace the legacy view licence page

When we first built our version of the view licence page's summary tab we applied the same logic when it came to a licence with multiple purposes; don't display abstraction amounts!

This is because the amounts could be different across the purposes and you can't just sum them. It is possible that the total amount of abstraction might be different on the licence when compared to the sum of its purposes (welcome to water abstraction!)

The old logic therefore hid the field. But we know anecdotally some users say you cannot see abstraction amounts in WRLS because of this behaviour.

So, in this change when a licence has multiple purposes we'll instead show a message and a link to the screen which shows the amounts for each purpose. Where a licence has one purpose, we'll continue to display the amounts.