Fix route handler property to follow same pattern
https://github.com/DEFRA/water-abstraction-team/issues/122

We have noticed that in some of our routes, the `handler` is within the `options property`.
```javascript
{
  method: 'POST',
  path: '/bills/{id}/remove',
  options: {
    handler: BillsController.submitRemove,
    auth: {
      access: {
        scope: ['billing']
      }
    }
  }
}
```

Whilst in others it is outside:
```javascript
{
  method: 'GET',
  path: '/bill-runs',
  handler: BillRunsController.index,
  options: {
    auth: {
      access: {
        scope: ['billing']
      }
    }
  }
}
```

This is as a result of some contraction between the [Hapi API documentation](https://hapi.dev/api/?v=21.3.3#-routeoptionshandler) having the `handler` within the `options` property,
and the [Hapi Tutorial](https://hapi.dev/tutorials/routing/?lang=en_US) giving examples with the `handler` outside the `options` property. The decision has been made to follow the API documentation.

This PR is focused on going through the routes files and ensuring the `handler` is within the `options` property.