Switch to GOV.UK table component for licence comms
https://eaflood.atlassian.net/browse/WATER-4315

> Part of the work to replace the legacy view licence page

This is a housekeeping task. When we rebuilt the tab we created the tables using the HTML details provided by the [design system](https://design-system.service.gov.uk/components/table/). Ideally, we use the Nunjucks component instead. We do this because it means when the package is updated the components will see those changes automatically. For the HTML, should something change, for example, a class is added, it is on us to manually update all places where we've used the HTML.

So, before this becomes fully live we are taking the opportunity to update the view to use the Nunjucks version of the component.