Fix bill run type in view licence bills tab
https://eaflood.atlassian.net/browse/WATER-4316

> Part of the work to replace the legacy view licence page

Whilst working on [Switch to GOV.UK table component for licence bills](https://github.com/DEFRA/water-abstraction-system/pull/1158) we spotted that a Nunjucks filter was being used to title-case the bill run type. How the bill run type is displayed is not as simple as just title-casing what the DB returns. In fact, because there is some logic involved and we have had to use it in a number of places we've added it to the base presenter as a reusable formatter.

This change corrects the view licence bills presenter to use the existing bill run type formatter.