Refactor presenter
We started with the aim of just updating the presenter to use the existing bill run type formatter. But we then spotted that the unit tests didn't follow our standard pattern of setting up test conditions in the `beforeEach()` (it was doing it in the `it`)

It was also testing things that are using existing formatters that have their own tests which means there is no need to repeat them here.

Finally, we felt we could simplify things if we renamed things to just be as they are (which meant some tweaks to the view as well). We also dropped `legacyId` from the presenter result because nothing in the view is using it.