Fix returns link in view licence page
https://eaflood.atlassian.net/browse/WATER-4557

> Part of the work to replace the legacy view licence page

When building our version of the returns tab in the view licence page we didn't spot that the link for 'complete' licences goes to a different URL than the one for 'due' licences.

Our current implementation is sending users to as page where we are asking them to enter the details again for a complete return!

This change fixes the issue.